Preliminary alpha complete
Talliliuqta mode is essentially complete. The computer will take cards if it notices that a set is available. If a nineth card is drawn, the field will be cleared and sent back into the deck. Clicking on the blank deck field once all cards are drawn will reset the game. AI may need to be scaled before increasing the difficulty.