Updating function to check_card_10
Developing an expandable function that will allow the computer to find a target value building up. The function should decrease the lowest index in the array by two, then check to see if the value in subsequent indices can be increased by one.

If the value is zero, increase along the even numbered indices (2, 4, etc)

if odd, increase along the odd numbered indices (3, 5, etc)

If the above fails for the range of the array, increase the array length by 1 and start over.