Built more scaleable card check function
A new card check function has been created that should be more scaleable. It appears to work as expected when used with a goal of 5 (Talliliuqta). The code should be useable for a goal of 10 (Quliliqta) once get check array cases for 7, 8, 9 and 10 are added. A goal of 20 should also be possible once all of the test cases from 10 to 20 are added as well