Fixed bug in Check Card
Fixed a bug in the check card function. Symmetrical values counted the array length without taking the card at the end of the array, causing cards to be counted twice. Added code to subtract one after trimming the array, signifying that there is one card less after shortening to the biggest card number.