Added keyboard/controller support, sound
Added controls for keyboard and controller. Rearranged all controls to be centered in a controller object instead of within objects. May have some glitches associated with this.

Added the ability to play sounds. Need to add the sound files for all of the different numbers in all of the different dialects